reestructurar env segons norma
