+++ include/environment_utils.h
@@ -1,22 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtins.h                                         :+:      :+:    :+:   */
+/*   environment_utils.h                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ecoma-ba <ecoma-ba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/02 13:02:56 by smercado          #+#    #+#             */
-/*   Updated: 2025/01/04 15:43:36 by ecoma-ba         ###   ########.fr       */
+/*   Created: 2025/01/04 15:42:14 by ecoma-ba          #+#    #+#             */
+/*   Updated: 2025/01/04 15:42:45 by ecoma-ba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef BUILTINS_H
-# define BUILTINS_H
+#ifndef ENVIRONMENT_UTILS_H
+# define ENVIRONMENT_UTILS_H
 # include "minishell.h"
 
-int		run_cd(t_command *command, t_environment *env);
-int		run_echo(t_command *command);
-int		run_pwd(t_command *command, t_environment *env);
-void	print_env(t_environment *env);
-
+void	update_env_value(t_environment *env, char *name, char *value);
+void	free_env_entry(t_environment *env);
 #endif
