+++ src/builtins/env.c
@@ -1,22 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtins.h                                         :+:      :+:    :+:   */
+/*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ecoma-ba <ecoma-ba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/02 13:02:56 by smercado          #+#    #+#             */
-/*   Updated: 2025/01/04 15:43:36 by ecoma-ba         ###   ########.fr       */
+/*   Created: 2025/01/04 15:37:26 by ecoma-ba          #+#    #+#             */
+/*   Updated: 2025/01/04 15:37:27 by ecoma-ba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef BUILTINS_H
-# define BUILTINS_H
-# include "minishell.h"
+#include "environment.h"
 
-int		run_cd(t_command *command, t_environment *env);
-int		run_echo(t_command *command);
-int		run_pwd(t_command *command, t_environment *env);
-void	print_env(t_environment *env);
-
-#endif
+/**
+ * Prints the whole environment.
+ */
+void	print_env(t_environment *env)
+{
+	while (env)
+	{
+		printf("%s=%s\n", env->name, env->value);
+		env = env->next;
+	}
+}
